+++ yk_utils/models/__init__.py
@@ -2,3 +2,4 @@
 from .deserialization import deserialize_model, deserialize_dbmodel
 from .base_model import Model
 from .error import ErrorSchema, Error
+from .base_enum import BaseEnum